requirements.txt
# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Image Processing
opencv-python>=4.8.0
Pillow>=9.5.0

# Data Handling & Metrics
numpy>=1.23.0
scikit-learn>=1.3.0
tqdm>=4.65.0

# Visualization (used by Grad-CAM heatmaps)
matplotlib>=3.7.0

--> kagglehub (unless you are downloading dataset locally)
--> flask, streamlit (unless app.py is used)